Back to kdeconnector file.
